refactor(transactions): collapse posting declarations to one type (p4)
Introduce a dependency-free leaf module `lib/transactions/posting.ts`
holding the canonical `Annotation` + `Posting` types. Point parser.ts
(`ParsedPosting = Posting`, re-export `Annotation`), model.ts (import +
re-export `Posting`), transactionRow.ts (`postings: Posting[]`), and
carryAnnotations.util.ts at it. Eliminates three independent structural
declarations of the same posting shape; `DraftPosting` already aliased
`Posting` from P2. The leaf also prevents a potential import cycle
between parser and model. No runtime changes. Mark REVIEW.md G3 done.
